+++ utils.c
@@ -31,6 +31,11 @@
 #include "mach_gettime.h"
 #endif
 
+#define DEFAULT_SERVER_INSTANCE	1
+#define TEN_TO_THE_POW_OF_THIRTEEN 10000000000000
+
+static uint instance = DEFAULT_SERVER_INSTANCE;
+
 gint64 janus_get_monotonic_time(void) {
 	struct timespec ts;
 	clock_gettime (CLOCK_MONOTONIC, &ts);
@@ -80,6 +85,10 @@ guint32 janus_random_uint32(void) {
 	return ret;
 }
 
+void janus_init_server_instance(uint server_instance) {
+	instance = server_instance;
+}
+
 guint64 janus_random_uint64_full(void) {
 	guint64 ret = 0;
 	if(RAND_bytes((void *)&ret, sizeof(ret)) != 1) {
@@ -90,7 +99,24 @@ guint64 janus_random_uint64_full(void) {
 }
 
 guint64 janus_random_uint64(void) {
-	return janus_random_uint64_full() & 0x1FFFFFFFFFFFFF;
+	/*
+	 * FIXME This needs to be improved, and use something that generates
+	 * more strongly random stuff... using /dev/urandom is probably not
+	 * a good idea, as we don't want to make it harder to cross compile Janus
+	 *
+	 * TODO Look into what libssl and/or libcrypto provide in that respect
+	 *
+	 * PS: JavaScript only supports integer up to 2^53, so we need to
+	 * make sure the number is below 9007199254740992 for safety
+	 */
+	/*
+	guint64 num = g_random_int() & 0x1FFFFF;
+	num = (num << 32) | g_random_int();
+	*/
+	guint64 num = g_random_int() & 0x8FF;
+	num = (instance * TEN_TO_THE_POW_OF_THIRTEEN) | (num << 32) | g_random_int();
+
+	return num;
 }
 
 char *janus_random_uuid(void) {
@@ -103,8 +129,8 @@ char *janus_random_uuid(void) {
 	const char *template = "xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx";
 	const char *samples = "0123456789abcdef";
 	union { unsigned char b[16]; uint64_t word[2]; } rnd;
-	rnd.word[0] = janus_random_uint64_full();
-	rnd.word[1] = janus_random_uint64_full();
+	rnd.word[0] = janus_random_uint64();
+	rnd.word[1] = janus_random_uint64();
 	/* Generate the string */
 	char uuid[37], *dst = uuid;
 	const char *p = template;